+++ crates/hc_sandbox/src/run.rs
@@ -1,13 +1,17 @@
 //! Helpers for running the conductor.
 
-use anyhow::anyhow;
 use std::path::Path;
 use std::process::Stdio;
 
+use anyhow::anyhow;
 use holochain_conductor_api::conductor::paths::ConfigFilePath;
 use holochain_conductor_api::conductor::paths::ConfigRootPath;
 use holochain_conductor_api::conductor::paths::KeystorePath;
 use holochain_conductor_api::conductor::{ConductorConfig, KeystoreConfig};
+use holochain_conductor_config::config::create_config;
+use holochain_conductor_config::config::read_config;
+use holochain_conductor_config::config::write_config;
+use holochain_conductor_config::ports::set_admin_port;
 use holochain_trace::Output;
 use holochain_types::websocket::AllowedOrigins;
 use tokio::io::AsyncBufReadExt;
@@ -18,9 +22,6 @@ use tokio::sync::oneshot;
 use crate::calls::attach_app_interface;
 use crate::calls::AddAppWs;
 use crate::cli::LaunchInfo;
-use crate::config::*;
-use crate::ports::random_admin_port;
-use crate::ports::set_admin_port;
 use crate::CmdRunner;
 
 // MAYBE: Export these strings from their respective repos
@@ -104,7 +105,7 @@ pub async fn run_async(
         Some(c) => c,
         None => {
             let passphrase = holochain_util::pw::pw_get()?;
-            let con_url = crate::generate::init_lair(
+            let con_url = holochain_conductor_config::generate::init_lair(
                 &config_root_path.is_also_data_root_path().try_into()?,
                 passphrase,
             )?;
@@ -115,7 +116,7 @@ pub async fn run_async(
         Some(port) => {
             set_admin_port(&mut config, port);
         }
-        None => random_admin_port(&mut config),
+        None => set_admin_port(&mut config, 0),
     }
     let _config_file_path = write_config(config_root_path.clone(), &config);
     let (tx_config, rx_config) = oneshot::channel();